requirements.txt
djangorestframework==3.11.0
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-webpack-loader==0.7.0
